Move MouseZoneManager to browser
